+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   [#228](https://github.com/cadeyrn/enterprise-policy-generator/issues/228)
 - Added support for a new policy type, needed for the updated OverrideFirstRunPage policy, see
   [#270](https://github.com/cadeyrn/enterprise-policy-generator/issues/270)
+- Added support for array fields in array objects, needed for the new AutoLaunchProtocolsFromOrigins policy, see
+  [#229](https://github.com/cadeyrn/enterprise-policy-generator/issues/229)
 - Added option for URL fields to require secure connection (https://), see
   [#277](https://github.com/cadeyrn/enterprise-policy-generator/issues/277)
 - Added option to set info link for array properties, see